requirements.txt
aiogram==3.13.1
fastapi==0.115.4
python-dotenv==1.0.1
pydantic==2.9.2
psycopg2==2.9.10
watchdog==5.0.3
PyYAML==6.0.2
jinja2==3.1.4
aiohttp==3.10.10
redis==5.2.0